Move search helpers into plugin e2e checks project
