+++ projects/plugins/jetpack/changelog/e2e-move-search-helpers
@@ -1,0 +1,4 @@
+Significance: minor
+Type: other
+
+E2E tests: move search helpers from e2e-commons to plugin e2e checks project
