+++ projects/plugins/jetpack/tests/e2e/helpers/search-helper.js
@@ -1,7 +1,8 @@
 import fs from 'fs';
 import path from 'path';
-import { execWpCommand } from './utils-helper';
+import { execWpCommand } from 'jetpack-e2e-commons/helpers/utils-helper';
 import config from 'config';
+import logger from 'jetpack-e2e-commons/logger';
 
 export async function enableInstantSearch() {
 	return execWpCommand( 'option update instant_search_enabled 1' );
@@ -102,6 +103,79 @@ function getSidebarsWidgetsData() {
 		'jetpack-instant-search-sidebar': [ 'jetpack-search-filters-8' ],
 		array_version: 3,
 	};
+}
+
+/**
+ * The function intercepts requests made to WPCOM Search API and returns mocked
+ * results to the frontend. It also simulates sorting and filtering.
+ *
+ * The route returns `searchResultForTest1` for query `test1`.
+ * And returns `searchResultForTest2` for any other queries.
+ *
+ * NOTE: The route sometimes is not persisted after page reloads so would need to
+ * call the function again to make sure.
+ *
+ * @param { Object } page - instance of a Playwright Page type
+ * @see https://playwright.dev/docs/api/class-page#pagerouteurl-handler
+ */
+export async function searchAPIRoute( page ) {
+	const SEARCH_API_PATTERN = /^https:\/\/public-api\.wordpress.com\/rest\/v1.3\/sites\/\d+\/search.*/;
+
+	await page.route( SEARCH_API_PATTERN, ( route, request ) => {
+		logger.info( `intercepted search API call: ${ request.url() }` );
+		const url = new URL( request.url() );
+		const params = url.searchParams;
+
+		// loads response for queries
+		let body;
+		switch ( params.get( 'query' ) ) {
+			case 'test1':
+				body = { ...searchResultForTest1 };
+				break;
+			case 'test2':
+			default:
+				body = { ...searchResultForTest2 };
+				break;
+		}
+
+		// deal with sorting
+		switch ( params.get( 'sort' ) ) {
+			case 'date_asc':
+				// put record 2 first
+				const tmpResult1 = body.results[ 0 ];
+				body.results[ 0 ] = body.results[ 1 ];
+				body.results[ 1 ] = tmpResult1;
+				break;
+			case 'date_desc':
+				// put record 3 first
+				const tmpResult2 = body.results[ 0 ];
+				body.results[ 0 ] = body.results[ 2 ];
+				body.results[ 2 ] = tmpResult2;
+				break;
+			case 'score_default':
+			default:
+				// the original sorting
+				break;
+		}
+
+		// deal with filtering: only works with one category and one tag by filtering the results array
+		const category = params.get( 'filter[bool][must][0][term][category.slug]' );
+		const tag = params.get( 'filter[bool][must][0][term][tag.slug]' );
+
+		if ( category ) {
+			body.results = body.results.filter( v => v?.categories?.includes( category ) );
+		}
+
+		if ( tag ) {
+			body.results = body.results.filter( v => v?.tags?.includes( tag ) );
+		}
+
+		route.fulfill( {
+			content: 'application/json',
+			headers: { 'Access-Control-Allow-Origin': '*' },
+			body: JSON.stringify( body ),
+		} );
+	} );
 }
 
 export const searchResultForTest1 = {
